Added a "seed" parameter, and made all the images be reproducible from it.
The GrowthImageBuilder owns the seed, and creates the Rng when
building, then passes to the GrowthImage.  If no seed is specified,
generate from OS random.

Palette generation is now done during the build() step, and has a
trait Palette for generating it, rather than taking Vec<RGB> at input.
The Rng from the GrowthImageBuilder is exposed to the Palette
implementations, so the palette generation is based on the seed as
well.